+++ tests/unit/test_microanalytic_score.py
@@ -33,3 +33,5 @@ def test_create_python_module():
         assert 'text/x-python' == json['type']
         assert 'public' == json['scope']
 
+
+